# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Soundwave Music Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  clerkId: string;
  username: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: Date;
}