# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Soundwave Music Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  /** @minLength 1 */
  username?: string;
  displayName?: string;
}